Android - Gastón - Add offer detail model.

+++ ButtonIntegration/ios/OfferDetail.h
@@ -21,7 +21,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 - (instancetype)initWithURL:(NSString *)url
                     offerID:(NSString *)offerID
-                       rate:(NSUInteger)rate
+                       rate:(double)rate
                 rateIsFixed:(BOOL)isFixed
           offerCreativeType:(NSString *)type;
